third_party/blink/web_tests/fast/events/context-nodrag-expected.txt

This tests to make sure that right clicking when the left-mouse button is pressed disables the drag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getSelection().isCollapsed is true
PASS successfullyParsed is true

TEST COMPLETE

This is a test box. ABC 123
